Got an i386/textmode.h driver working, implemented kernel logging.

+++ include/nop/start.h
@@ -16,6 +16,6 @@ struct start_task_t {
   void (*func)(void);
 };
 
-void start(start_block_t *blocks, int block_count, start_task_t *tasks, int task_count);
+void start(start_block_t *blocks, size_t block_count, start_task_t *tasks, size_t task_count);
 
 #endif
